requirements.txt
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
opencv-python==4.8.1.78
PyYAML==6.0.1
tqdm==4.66.1
watchdog==2.1.9
matplotlib
scipy==1.11.4
timm